Renaming getCSV to getCSVData for clarity
