+++ lib/list.js
@@ -152,7 +152,7 @@ module.exports = function (keystone) {
 	List.prototype.field = require('./list/field');
 	List.prototype.get = List.prototype.set = require('./list/set');
 	List.prototype.getAdminURL = require('./list/getAdminURL');
-	List.prototype.getCSV = require('./list/getCSV');
+	List.prototype.getCSVData = require('./list/getCSVData');
 	List.prototype.getData = require('./list/getData');
 	List.prototype.getDocumentName = require('./list/getDocumentName');
 	List.prototype.getOptions = require('./list/getOptions');
